rename some empty illustrations
