+++ ui/src/components/dashboard/components/ChartViewWrapper.vue
@@ -26,7 +26,7 @@
 
 <script lang="ts" setup>
     import Sections from "../sections/Sections.vue"
-    import EmptyVisualDashboard from "../../../assets/empty_visuals/Visuals_empty_dashboard.svg"
+    import EmptyVisualDashboard from "../../../assets/empty_visuals/dashboard.svg"
     import {useDashboardStore} from "../../../stores/dashboard"
 
     const dashboardStore = useDashboardStore()
